refactor: extract vinext:fix-use-server-closure-collision into its own plugin file
Move the inline plugin object from index.ts into
packages/vinext/src/plugins/fix-use-server-closure-collision.ts,
following the same pattern as async-hooks-stub, client-reference-dedup,
and optimize-imports. index.ts now imports and references the exported
fixUseServerClosureCollisionPlugin constant. No behaviour change.
